Добавил еще немного логики

+++ src/main/java/com/samuilolegovich/WBL/model/Bets.java
@@ -1,5 +1,5 @@
 package com.samuilolegovich.WBL.model;
 
 public interface Bets {
-    public Win calculateTheWin(long id, long bet, RedBlack redBlack);
+    public Win calculateTheWin(long id, int bet, RedBlack redBlack);
 }
